requirements.txt
fastapi
langchain
langchain-community
langsmith
pydantic
python-dotenv
faiss-cpu
youtube-transcript-api
streamlit
langserve
uvicorn
sse_starlette